frameworks/django/testsite/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
mysqlclient==2.2.4
sqlparse==0.4.4
